YLP-2783 remove upload id references (#131)
* wip: implementing upload id removal

* update changelog

* remove arm to fix publish

* restore go.sum...

* implement featur flag

* update changelogs

* fix BLP tests and bump rc2 version

* bump version

+++ data/service/api/v1/datasets_delete.go
@@ -31,7 +31,6 @@ type dataSetsDeleteParams struct {
 // @Router /v1/datasets/:dataSetID [delete]
 func DataSetsDelete(dataServiceContext dataService.Context) {
 	req := dataServiceContext.Request()
-	ctx := req.Context()
 
 	dataSetID := req.PathParam("dataSetId")
 	if dataSetID == "" {
@@ -39,47 +38,50 @@ func DataSetsDelete(dataServiceContext dataService.Context) {
 		return
 	}
 
-	dataSet, err := dataServiceContext.DataRepository().GetDataSetByID(ctx, dataSetID)
-	if err != nil {
-		dataServiceContext.RespondWithInternalServerFailure("Unable to get data set by id", err)
-		return
-	}
-	if dataSet == nil {
-		// FIXME: This is a temporary fix, it should return an error.
-		dataServiceContext.RespondWithStatusAndData(http.StatusOK, ErrorDataSetIDNotFound(dataSetID))
-		return
-	}
-
-	targetUserID := dataSet.UserID
-	if targetUserID == nil || *targetUserID == "" {
-		dataServiceContext.RespondWithInternalServerFailure("Unable to get user id from data set")
-		return
-	}
-
-	permissions, err := dataServiceContext.PermissionClient().GetUserPermissions(req, *targetUserID)
-	if err != nil {
-		if request.IsErrorUnauthorized(err) {
-			dataServiceContext.RespondWithError(service.ErrorUnauthorized())
-		} else {
-			dataServiceContext.RespondWithInternalServerFailure("Unable to get user permissions", err)
+	if dataServiceContext.IsUploadIdUsed() {
+		ctx := req.Context()
+		dataSet, err := dataServiceContext.DataRepository().GetDataSetByID(ctx, dataSetID)
+		if err != nil {
+			dataServiceContext.RespondWithInternalServerFailure("Unable to get data set by id", err)
+			return
 		}
-		return
-	}
-	if !permissions {
-		dataServiceContext.RespondWithError(service.ErrorUnauthorized())
-		return
-	}
+		if dataSet == nil {
+			// FIXME: This is a temporary fix, it should return an error.
+			dataServiceContext.RespondWithStatusAndData(http.StatusOK, ErrorDataSetIDNotFound(dataSetID))
+			return
+		}
 
-	// Read delete options (remove dataset entry ?):
-	var jsonParams dataSetsDeleteParams
-	doPurge := false
-	if err := req.DecodeJsonPayload(&jsonParams); err == nil {
-		doPurge = jsonParams.Purge
-	}
+		targetUserID := dataSet.UserID
+		if targetUserID == nil || *targetUserID == "" {
+			dataServiceContext.RespondWithInternalServerFailure("Unable to get user id from data set")
+			return
+		}
 
-	if err = dataServiceContext.DataRepository().DeleteDataSet(ctx, dataSet, doPurge); err != nil {
-		dataServiceContext.RespondWithInternalServerFailure("Unable to delete data", err)
-		return
+		permissions, err := dataServiceContext.PermissionClient().GetUserPermissions(req, *targetUserID)
+		if err != nil {
+			if request.IsErrorUnauthorized(err) {
+				dataServiceContext.RespondWithError(service.ErrorUnauthorized())
+			} else {
+				dataServiceContext.RespondWithInternalServerFailure("Unable to get user permissions", err)
+			}
+			return
+		}
+		if !permissions {
+			dataServiceContext.RespondWithError(service.ErrorUnauthorized())
+			return
+		}
+
+		// Read delete options (remove dataset entry ?):
+		var jsonParams dataSetsDeleteParams
+		doPurge := false
+		if err := req.DecodeJsonPayload(&jsonParams); err == nil {
+			doPurge = jsonParams.Purge
+		}
+
+		if err = dataServiceContext.DataRepository().DeleteDataSet(ctx, dataSet, doPurge); err != nil {
+			dataServiceContext.RespondWithInternalServerFailure("Unable to delete data", err)
+			return
+		}
 	}
 
 	dataServiceContext.RespondWithStatusAndData(http.StatusOK, struct{}{})
